add "printhex" mode and functionality to program
This module adds a header with the 'readfile' function that takes std::string filePath and size_t neededLength as the parameters and outputs the std::vector<uint8_t> with needed number of bytes from file in filePath. Also, it integrates it into the mode reception program and gets parameters through ReadFileParams structure, if mode 'printhex' is chosen and the correct number of arguments is given.

Also, changes to .gitignore to not track test resources for now. Changes to CMakeLists to change executable name, state C++11 standard and use new source files for readfile.