refactor main.cpp and sound-effects + add all new headers to CMakeLists
This update renames a couple of mode functions and parameter structures but most importantly, it changes how the mode functions get parameters. Now the parameter structure process and store POSIX-like options from the command arguments and pass them to the place they need be used in. The effect modes are now done by a single function called run_mode_effect() which is a template that calls different effect functions depending on the type of input options structure T. In general, the argument parsing became more organized but there's still room for improvement.